fix: move params out of keys
also rename oauth sub to provider from app
