+++ .github/workflows/release.yml
@@ -127,7 +127,7 @@ jobs:
       - name: OpenApi
         uses: readmeio/rdme@8.6.6
         with:
-          rdme: openapi openapi.yaml --key=${{ secrets.README_API_KEY }} --id=65566875eee211003a3b3a47
+          rdme: openapi openapi.yaml --key=${{ secrets.README_API_KEY }} --id=655bf19aed925600124be104
 
       - name: Release
         uses: ncipollo/release-action@v1
